+++ package/core/src/components/form/TextInput/TextInput.tsx
@@ -23,11 +23,12 @@ const CustomTextInput: FormProps<CustomTextInputProps> = (props) => {
           color: getColor(color ?? defaultProps.color ?? "global"),
           fontSize: GetTextSizeValue(
             fontSize ?? defaultProps.fontSize,
-            theme.fontSizes.text
+            theme.fontSizes.text,
           ),
+          height: "100%",
         }}
         placeholderTextColor={getColor(
-          placeholderTextColor ?? defaultProps.placeholderTextColor ?? "muted"
+          placeholderTextColor ?? defaultProps.placeholderTextColor ?? "muted",
         )}
         {...textInputProps}
       />